[tech debt] Create custom hook for movie fetch
